pyproject.toml
[tool.poetry]
name = "discord-bot-sample"
version = "0.1.0"
description = ""
authors = ["taaaaho <taaaaho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
